Changed the way movement works. Direction is now for the entire snake (altering only the first section) and all other sections now follow the first one. Still need to correct upwards movement and for some reason pair segments are disappearing

+++ Drawable.hpp
@@ -11,10 +11,10 @@ namespace SnakeGame {
  */
 struct Drawable {
 	virtual void draw(Screen & screen) = 0;
-	double m_x;
-	double m_y;
+	int m_x;
+	int m_y;
 
-	Drawable(double x, double y): m_x(x), m_y(y) {}
+	Drawable(int x, int y): m_x(x), m_y(y) {}
 };
 
 } // namespace SnakeGame
